+++ src/Uno.UI/UI/Xaml/Controls/MenuFlyout/MenuFlyoutPresenter.Properties.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.UI.Xaml.Controls.Primitives;
+
+namespace Microsoft.UI.Xaml.Controls;
+
+partial class MenuFlyoutPresenter
+{
+	public MenuFlyoutPresenterTemplateSettings TemplateSettings { get; private set; }
+}
